+++ Runtime/Model/BacktraceData.cs
@@ -45,7 +45,7 @@ namespace Backtrace.Unity.Model
         /// <summary>
         /// Version of the C# library
         /// </summary>
-        public const string AgentVersion = "3.0.2";
+        public const string AgentVersion = "3.0.3";
 
         /// <summary>
         /// Application thread details