+++ CHANGELOG.md
@@ -29,7 +29,7 @@ For example, to only get error reporting for hangs or crashes then only return f
   * Database add operation time (`performance.database`),
   * Database single send method time (`performance.send`),
   * Database single flush method time (`performance.flush`)
-  
+- Improvements to JIT stack frame parsing.
 
 ## Version 3.0.2
 - `BacktraceDatabase` now provides a new `Send` method. This method will try to send all objects from the database respecting the client side deduplication and retry setting. This can be used as an alternative to the `Flush` method which will try to send all objects from the database ignoring any client side deduplication and retry settings.